+++ packages/desktop/app/utils/survey.js
@@ -1,7 +1,7 @@
 // Much of this file is duplicated in `packages/mobile/App/ui/components/Forms/SurveyForm/helpers.ts`
 import React from 'react';
 import * as yup from 'yup';
-import { checkJSONCriteria } from '@tamanu/shared/utils';
+import { checkJSONCriteria } from '@tamanu/shared/utils/criteria';
 import { intervalToDuration, parseISO } from 'date-fns';
 
 import {
